flow: add handlers for structs to netlink
This adds handers for structs to netlinnk. These are useful
in flow when doing sets but also when using flow as an IPC
which is useful for debugging. And working with user mode
drivers while they are being ported into the kernel.

Plus a fix for PID. Note I'll need to smash all these
commit into something coherent before submitting anyways

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>