flow: fixup headers
if_flow.h in UAPI is not much smaller only including the necessary
bits for the messaging protocol. We do not want to expose internal
data structures of the kernel. Its much better to only expose the
minimal set of data structures required and let the others evolve
indpendently.

The downside is we need to duplicate structures here.

Finally flow.c and flowlib.c need to be updated for new header
formats.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>