flow: use instance id in match arguments instead of header id
The header ID may not be unique in the case where a packet may
have multiple stacked headers. For example matching on a VLAN
header is unclear for a Q in Q packet.

Same goes for tunnel packets. Its unclear which UDP header to
match for example when UDP is used for tunnel encapsolation

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>