flow: off by one error in find_fields do an explicit uid return
fields are index'ed from 0 in null terminated field
array but uid's start at 1 so we end up returning the wrong
uid value when simply returning 'i'. Also it assumes that the
array index and uid match which is not necessarily true and
has only been true so far because the flow pipeline is generated
this way.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>