flow: add new attribute apply_action_group
Currently we use the source boundary to indicate when actions
are applied. When a source boundary is crossed all previous
actions should be applied. However we are currently also using
it to "create" tables inside a source id.

This means its overloaded and confusing. This patch cleans up
the source id and adds a new attr called apply_action_group
which can be used to learn when actions are applied.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>