flow: add support for better_pipeline
Better pipelin is a fictional pipeline used to test the API. It
includes a set of tables and a TCAM that can be used to create
and destroy additional tables.

These were just made up to test some functionality of the API
its not intended to represent any real devices. I'll be adding
a set of other models shortly to test other configurations.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>