flow: add a trap action
I'm thinking about how to model a cpu port in the switch/rocker
environment. Adding a CPU trap action illustrates the problem.

Where do trapped packets show up? My initial response is on a
netdev with a feature bit set indicating it is into the switch
pipeline and not a switch port.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>